[Chore] #7 - 코드리뷰 반영
